requirements.txt
# Agentic Program Dependencies
ollama>=0.4.0
gradio>=4.0.0
playwright>=1.40.0
python-docx>=1.1.0
pyautogui>=0.9.54
pywin32>=306
Pillow>=10.0.0
